teeny weeny preprocessing fixes
